server-django-app/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.3
channels==2.4.0
channels-redis==3.0.1
constantly==15.1.0
cryptography==3.2.1
daphne==2.5.0
Django==2.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
iniconfig==1.1.1
msgpack==1.0.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.2
pytest-asyncio==0.14.0
pytest-django==3.9.0
pytz==2020.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.2.0
